tekrar 1

+++ src/main/java/com/uur/Authentications/JpaRepositories/AuthorityRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 
 @RepositoryRestResource(  exported = false)
 @Repository
-public interface AuthorityRepository extends JpaRepository<Authority, Integer> {
+public interface AuthorityRepository extends JpaRepository<Authority, Long> {
 
     Optional<Authority> findByName(String name);
 }
