+++ src/main/java/com/uur/Authentications/dtos/UserRoleDto.java
@@ -6,8 +6,8 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class UserRoleDto {
-    private int id;
-    private int roleId;
+    private long id;
+    private long roleId;
     private String roleName;
-    private int userId;
+    private long userId;
 }
